make payment summary interactive
