+++ checkout.html
@@ -82,6 +82,6 @@
         </div>
       </div>
     </div>
-    <script type="module" src="/scripts/checkout.js"></script>
+    <script type="module" src="/scripts/checkout/orderSummary.js"></script>
   </body>
 </html>
