+++ scripts/utils/requiredFunctions.js
@@ -1,3 +1,3 @@
 export function fixmoney(money) {
-    return (money / 100 .toFixed(2)) ;
+    return (Math.round(money) / 100 .toFixed(2)) ;
 }
